requirements.txt
streamlit
fpdf
python-dotenv
openai